combined indexing and filtering

+++ src/delete_files.rs
@@ -1,21 +1,21 @@
 use std::{path::PathBuf, fs, io};
 use filesize::file_real_size;
 
-pub fn delete_files(filtered_files: Vec<PathBuf>) {
-    // amount of files = length of our vector
-    let file_amount = filtered_files.len();
+pub fn delete_files(index: (Vec<PathBuf>, usize)) {
     let mut total_file_size = 0;
 
     // iterates through filtered files, adds size to total, then deletes them
-    for file in filtered_files.iter() {
+    for file in index.0.iter() {
         total_file_size += file_real_size(file).unwrap();
-        fs::remove_file(file).unwrap();
+        if !file.as_path().to_str().unwrap().contains(&"song") {
+            fs::remove_file(file).unwrap();
+        }
     }
-    println!("\n{} files deleted totaling {}mb", file_amount, total_file_size/1000000);
+    println!("\n{} files deleted totaling {}mb", index.1, total_file_size/1000000);
     
     // idk how to make the program not just immediately close itself
-    let mut useless = String::new();
+    let mut wait = String::new();
     io::stdin()
-        .read_line(&mut useless)
+        .read_line(&mut wait)
         .expect("Invalid input.");
 }
